Use FormRequests for passkey bridge endpoints and add German translations
