+++ lang/de.json
@@ -373,6 +373,7 @@
     "Answer": "Antworten",
     "Answer to customer": "Antwort an Kunden",
     "API Port": "API Port",
+    "App bridge unavailable.": "App-Bridge nicht verfügbar.",
     "App Push": "App Push",
     "Append links": "Links hinzufügen",
     "Applies to Departments": "Gilt für Abteilungen",
@@ -426,6 +427,7 @@
     "Authenticatable Id": "Authentifizierbare ID",
     "Authenticatable Type": "Authentifizierbarer Typ",
     "authenticatable.name": "Name",
+    "Authenticate with your passkey": "Mit Passkey anmelden",
     "Authenticator app": "Authenticator-App",
     "Author": "Autor",
     "Authorisation user": "Freigabeberechtigter",
@@ -572,6 +574,7 @@
     "checkout-finish.get": "Bestellung abschließen",
     "checkout.get": "Zur Kasse",
     "Children": "Untergeordnete",
+    "Choose a name for this device, then confirm with biometrics.": "Wähle einen Namen für dieses Gerät und bestätige dann mit Biometrie.",
     "Choose bank account": "Bankkonto auswählen",
     "Choose range": "Bereich auswählen",
     "City": "Ort",
@@ -644,6 +647,8 @@
     "Confirm Cancellation": "Kündigung bestätigen",
     "Confirm Delete Event": "Termin löschen",
     "Confirm Option": "Bestätigungsoption",
+    "Confirm with biometrics in the system browser.": "Bestätige mit Biometrie im System-Browser.",
+    "Confirm with biometrics…": "Mit Biometrie bestätigen…",
     "Confirmed": "Bestätigt",
     "Connection": "Verbindung",
     "Constraints": "Einschränkungen",
@@ -731,6 +736,7 @@
     "Create New Group": "Neue Gruppe erstellen",
     "Create Orders": "Aufträge erstellen",
     "Create Orders From Work Times": "Aufträge aus Arbeitszeiten erstellen",
+    "Create passkey": "Passkey erstellen",
     "Create Payment Reminder": "Zahlungserinnerung erstellen",
     "Create Payment Run": "Zahlungslauf erstellen",
     "Create Payment Run|Do you really want to create the Payment Run?|Cancel|Yes": "Zahlungslauf erstellen|Möchten Sie den Zahlungslauf wirklich erstellen?|Abbrechen|Ja",
@@ -904,6 +910,7 @@
     "Details": "Details",
     "Detecting document...": "Dokument wird erkannt...",
     "Device": "Gerät",
+    "Device name": "Gerätename",
     "Device Token": "Geräte-Token",
     "Device token deleted": "Geräte-Token gelöscht",
     "Device Tokens": "Geräte-Tokens",
@@ -1619,6 +1626,7 @@
     "Minutes": "Minuten",
     "minutes": "Minuten",
     "Mixed": "Gemischt",
+    "Mobile Device": "Mobiles Gerät",
     "Mobile Push Notifications": "Mobile Push-Benachrichtigungen",
     "Model Id": "Modell-ID",
     "Model Type": "Model",
@@ -1878,6 +1886,10 @@
     "partial_delivered": "teilweise geliefert",
     "partial_paid": "teilweise bezahlt",
     "Passkey": "Passkey",
+    "Passkey authentication failed.": "Passkey-Anmeldung fehlgeschlagen.",
+    "Passkey is not supported in this environment.": "Passkeys werden in dieser Umgebung nicht unterstützt.",
+    "Passkey Login": "Passkey-Anmeldung",
+    "Passkey registration failed.": "Passkey-Registrierung fehlgeschlagen.",
     "Passkeys": "Passkeys",
     "Password": "Passwort",
     "Password reset failed": "Passwort zurücksetzen fehlgeschlagen",
@@ -1976,6 +1988,7 @@
     "Pivot Id": "Pivot-ID",
     "Place Of Birth": "Geburtsort",
     "Platform": "Plattform",
+    "Please enter a name.": "Bitte gib einen Namen ein.",
     "Please request a new one.": "Bitte fordern Sie einen Neuen an.",
     "Please scan at least one document.": "Bitte mindestens ein Dokument scannen.",
     "Please select": "Bitte auswählen",
@@ -2181,8 +2194,10 @@
     "Refund Date": "Gutschriftsdatum",
     "Regards": "Mit freundlichen Grüßen",
     "Regions": "Regionen",
+    "Register a new passkey": "Neuen Passkey registrieren",
     "Register a passkey": "Passkey registrieren",
     "Register a passkey first": "Bitte zuerst einen Passkey registrieren",
+    "Register Passkey": "Passkey registrieren",
     "Reject": "Ablehnen",
     "Rejected": "Abgelehnt",
     "rejected": "Abgelehnt",
@@ -2467,6 +2482,7 @@
     "Signed Date": "Unterzeichnet am",
     "Signed Documents": "Unterzeichnete Dokumente",
     "Signed Sepa Mandate": "Unterzeichnetes Sepa Mandat",
+    "Signing in…": "Anmeldung läuft…",
     "Simulate": "Simulieren",
     "Single Booking": "Einzelbuchung",
     "Size": "Größe",
@@ -2526,6 +2542,7 @@
     "Stock will be reserved on creation of given documents.": "Bestand wird bei Erstellung der gegebenen Dokumente reserviert.",
     "Storage": "Speicher",
     "Stores Serial Numbers": "Speichert Seriennummern",
+    "Storing passkey…": "Passkey wird gespeichert…",
     "Street": "Straße",
     "street": "Straße",
     "Strikethrough": "Durchgestrichen",
@@ -2767,6 +2784,7 @@
     "Transactions": "Transaktionen",
     "Translation": "Übersetzung",
     "Translations": "Übersetzungen",
+    "Try again": "Erneut versuchen",
     "Try to match all open transactions with orders": "Alle offenen Transaktionen mit Aufträgen abgleichen",
     "Tue": "Di",
     "Tuesday": "Dienstag",
@@ -3113,6 +3131,7 @@
     "Your browser does not support preview for this file.": "Ihr Browser unterstützt keine Vorschau für diese Datei.",
     "your comment": "Ihren Kommentar",
     "your contact": "Ihren Kontakt",
+    "Your device should prompt for biometric verification.": "Dein Gerät sollte zur biometrischen Bestätigung auffordern.",
     "Your documents are ready for download.": "Ihre Dokumente sind zum Download bereit.",
     "Your documents have been created.": "Ihre Dokumente wurden erstellt.",
     "Your export is ready for download.": "Ihr Export ist zum Download bereit.",
