+++ src/Http/Requests/PasskeyBridgeExchangeRequest.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace FluxErp\Http\Requests;
+
+class PasskeyBridgeExchangeRequest extends BaseFormRequest
+{
+    public function rules(): array
+    {
+        return [
+            'code' => ['required', 'string', 'size:64'],
+            'code_verifier' => ['required', 'string', 'min:43', 'max:128'],
+        ];
+    }
+}
