+++ src/Http/Requests/PasskeyBridgeShowRegisterRequest.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace FluxErp\Http\Requests;
+
+class PasskeyBridgeShowRegisterRequest extends BaseFormRequest
+{
+    public function rules(): array
+    {
+        return [
+            'transfer_token' => ['required', 'string', 'size:64'],
+        ];
+    }
+}
